Refactor: TopNavigation 합성 컴포넌트로 수정
