+++ src/components/TopNavigation/TopNavigationTitle.tsx
@@ -1,0 +1,6 @@
+'use client';
+import { PropsWithChildren } from 'react';
+
+export const TopNavigationTitle = ({ children }: PropsWithChildren) => {
+  return <div className="flex w-1/3 justify-center">{children}</div>;
+};
